back-end/requirements.txt
aiohttp==3.5.4
aiohttp-autoreload==0.0.1
aiohttp-jinja2==1.1.1
alembic==1.0.10
aniso8601==7.0.0
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
bcrypt==3.1.6
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
Flask==1.0.3
Flask-Cors==3.0.8
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.11
MarkupSafe==1.1.1
multidict==4.5.2
psycopg2==2.8.2
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.4
Werkzeug==0.15.4
yarl==1.3.0
